// src/client/domains_gen/Task.ts
export type TaskDomain = Domain<TaskObject>;

export interface TaskObject {
    name: FieldRef<string>;
    k: FieldRef<string>;
    a: FieldRef<string>;
    b: FieldRef<string>;
    t: FieldRef<string>;
    children: KeyValue<TaskObject>;
    assignee: FieldRef<PersonObject | undefined>;
}

export type TaskObjectKV = KeyValue<unknown, TaskObject>;

// Generated by 30000029
export const TaskDomain = {
    id: "9564b712-4d9b-4bd8-88f2-0811037270b5",
    name: "Task",
    defaultObjectRoot: undefined as ObjectRoot | undefined,

    isMyInstance: (obj: any): obj is TaskObject => {
        return typeof obj === "object" && obj.domain === TaskDomain;
    },

    // Generated by 30000021
    getOrCreate: (id: ObjectId | undefined): TaskObject => {
        if (!TaskDomain.defaultObjectRoot)
            throw new Error(
                `CODE00000000 Task.getOrCreate You should first set defaultObjectRoot before using getOrCreate or else use getOrCreateEx!`,
            );
        return TaskDomain.getOrCreateEx(TaskDomain.defaultObjectRoot, id);
    },

    getOrCreateEx: (objectRoot: ObjectRoot, id: ObjectId | undefined): TaskObject => {
        if (id && id.split(":")[0] !== "9564b712-4d9b-4bd8-88f2-0811037270b5")
            throw new Error(`CODE00000000 Task.getOrCreate Incorrect domainId in id = ${id}`);
        if (!id) {
            id = "9564b712-4d9b-4bd8-88f2-0811037270b5:" + newId();
        }
        const r = new KeyValueToMsgTranslator<unknown, TaskObject>({
            baseTypeMeta: objectRoot.odbConnection.odbBase.__metaRoot.expectType("Task"),
            stateReader: objectRoot.odbConnection,
            actionProcessor: objectRoot.stdActionProcessor,
            path: ["Task", id],
            createIfNotExists: true,
        });
        return r.asObj();
    },

    // Generated by 30000021
    getOrThrow: (id: ObjectId): TaskObject => {
        if (!TaskDomain.defaultObjectRoot)
            throw new Error(`CODE00000000 Task.getOrThrow You should first set defaultObjectRoot before using getOrThrow or else use getOrThrowEx!`);
        return TaskDomain.getOrThrowEx(TaskDomain.defaultObjectRoot, id);
    },

    getOrThrowEx: (objectRoot: ObjectRoot, id: ObjectId): TaskObject => {
        if (id && id.split(":")[0] !== "9564b712-4d9b-4bd8-88f2-0811037270b5")
            throw new Error(`CODE00000000 Task.getOrThrow Incorrect domainId in id = ${id}`);
        if (!id) {
            throw new Error(`CODE00000000 Task.getOrThrow id is empty!`);
        }
        const r = new KeyValueToMsgTranslator<unknown, TaskObject>({
            baseTypeMeta: objectRoot.odbConnection.odbBase.__metaRoot.expectType("Task"),
            stateReader: objectRoot.odbConnection,
            actionProcessor: objectRoot.stdActionProcessor,
            path: ["Task", id],
            createIfNotExists: true,
        });
        return r.asObj();
    },

    // Generated by 30000021
    getOrUndefined: (id: ObjectId | undefined): TaskObject | undefined => {
        if (!TaskDomain.defaultObjectRoot)
            throw new Error(
                `CODE00000000 Task.getOrUndefined You should first set defaultObjectRoot before using getOrUndefined or else use getOrUndefinedEx!`,
            );
        return TaskDomain.getOrUndefinedEx(TaskDomain.defaultObjectRoot, id);
    },

    getOrUndefinedEx: (objectRoot: ObjectRoot, id: ObjectId | undefined): TaskObject | undefined => {
        if (id && id.split(":")[0] !== "9564b712-4d9b-4bd8-88f2-0811037270b5")
            throw new Error(`CODE00000000 Task.getOrUndefined Incorrect domainId in id = ${id}`);
        if (!id) {
            return undefined;
        }
        const r = new KeyValueToMsgTranslator<unknown, TaskObject>({
            baseTypeMeta: objectRoot.odbConnection.odbBase.__metaRoot.expectType("Task"),
            stateReader: objectRoot.odbConnection,
            actionProcessor: objectRoot.stdActionProcessor,
            path: ["Task", id],
            createIfNotExists: true,
        });
        return r.asObj();
    },

    // Generated by 30000025
    verifyPermissions: (objectRoot: ObjectRoot, object: TaskObject, domainAction: DomainActionSerialized): boolean => {
        return true;
    },

    // Generated by 30000050
    beforeSendAction: (
        objectRoot: ObjectRoot,
        object: TaskObject,
        domainAction: DomainActionSerialized,
        targetDataSide: DataSide,
        destSession: SessionObject | undefined,
    ): DomainActionSerialized | undefined => {
        let cloned: true | undefined;
        let change = domainAction;

        //-------------------------------------------- Custom code start --------------------------------------------

        //-------------------------------------------- Custom code end --------------------------------------------

        return change;
    },
};